Move RedirectURLForm from CustomDomainListScreen to EndpointDirectAccessScreen
